Address SonarCloud concerns
