--- robozonky-common/src/main/java/com/github/robozonky/common/tenant/TransactionalTenantImpl.java
@@ -1,114 +1,0 @@
-/*
- * Copyright 2018 The RoboZonky Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.robozonky.common.tenant;
-
-import java.util.Optional;
-import java.util.Queue;
-import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.function.Function;
-
-import com.github.robozonky.api.SessionInfo;
-import com.github.robozonky.api.remote.entities.Restrictions;
-import com.github.robozonky.api.strategies.InvestmentStrategy;
-import com.github.robozonky.api.strategies.PurchaseStrategy;
-import com.github.robozonky.api.strategies.SellStrategy;
-import com.github.robozonky.common.remote.Zonky;
-import com.github.robozonky.common.state.InstanceState;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * {@link #getState(Class)} returns {@link TransactionalInstanceState} instead of the default {@link InstanceState}
- * implementation. Every other method delegated to the default {@link Tenant} implementation.
- */
-final class TransactionalTenantImpl implements TransactionalTenant {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(TransactionalTenantImpl.class);
-    private final Tenant parent;
-    private final Queue<Runnable> stateUpdates = new ConcurrentLinkedQueue<>();
-
-    public TransactionalTenantImpl(final Tenant parent) {
-        this.parent = parent;
-    }
-
-    @Override
-    public <T> T call(final Function<Zonky, T> operation, final ZonkyScope scope) {
-        return parent.call(operation, scope);
-    }
-
-    @Override
-    public boolean isAvailable(final ZonkyScope scope) {
-        return parent.isAvailable(scope);
-    }
-
-    @Override
-    public RemotePortfolio getPortfolio() {
-        return parent.getPortfolio();
-    }
-
-    @Override
-    public Restrictions getRestrictions() {
-        return parent.getRestrictions();
-    }
-
-    @Override
-    public SessionInfo getSessionInfo() {
-        return parent.getSessionInfo();
-    }
-
-    @Override
-    public Optional<InvestmentStrategy> getInvestmentStrategy() {
-        return parent.getInvestmentStrategy();
-    }
-
-    @Override
-    public Optional<SellStrategy> getSellStrategy() {
-        return parent.getSellStrategy();
-    }
-
-    @Override
-    public Optional<PurchaseStrategy> getPurchaseStrategy() {
-        return parent.getPurchaseStrategy();
-    }
-
-    @Override
-    public <T> InstanceState<T> getState(final Class<T> clz) {
-        LOGGER.trace("Creating transactional instance state for {}.", clz);
-        return new TransactionalInstanceState<>(stateUpdates, parent.getState(clz));
-    }
-
-    @Override
-    public synchronized void close() throws Exception {
-        parent.close();
-        if (!stateUpdates.isEmpty()) {
-            throw new IllegalStateException("There are uncommitted changes.");
-        }
-    }
-
-    @Override
-    public synchronized void commit() {
-        LOGGER.debug("Replaying transaction.");
-        while (!stateUpdates.isEmpty()) {
-            stateUpdates.poll().run();
-        }
-    }
-
-    @Override
-    public synchronized void abort() {
-        stateUpdates.clear();
-    }
-}
